export mediatag component
