+++ src/screens/Feed/components/MediaTag/index.js
@@ -1,0 +1,1 @@
+export * from './component';
